+++ scrapy/shell.py
@@ -11,20 +11,20 @@ from w3lib.url import any_to_uri
 
 from scrapy.item import BaseItem
 from scrapy.spider import BaseSpider
-from scrapy.selector import XPathSelector, XmlXPathSelector, HtmlXPathSelector
+from scrapy.selector import Selector
 from scrapy.utils.spider import create_spider_for_request
 from scrapy.utils.misc import load_object
 from scrapy.utils.response import open_in_browser
 from scrapy.utils.console import start_python_console
 from scrapy.settings import Settings
-from scrapy.http import Request, Response, HtmlResponse, XmlResponse
+from scrapy.http import Request, Response
 from scrapy.exceptions import IgnoreRequest
 
 
 class Shell(object):
 
     relevant_classes = (BaseSpider, Request, Response, BaseItem,
-        XPathSelector, Settings)
+                        Selector, Settings)
 
     def __init__(self, crawler, update_vars=None, code=None):
         self.crawler = crawler
@@ -95,10 +95,7 @@ class Shell(object):
         self.vars['spider'] = spider
         self.vars['request'] = request
         self.vars['response'] = response
-        self.vars['xxs'] = XmlXPathSelector(response) \
-            if isinstance(response, XmlResponse) else None
-        self.vars['hxs'] = HtmlXPathSelector(response) \
-            if isinstance(response, HtmlResponse) else None
+        self.vars['sel'] = Selector(response)
         if self.inthread:
             self.vars['fetch'] = self.fetch
         self.vars['view'] = open_in_browser